# swagger.yaml
openapi: 3.0.1
info:
  title: PlantMD
  description: This is the documentation for the backend API supporting PlantMD.
  contact:
    email: rachitb@gatech.edu
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://crincs-dot-thinger.ue.r.appspot.com
  description: Google Cloud temporary URL
tags:
- name: disease
  description: Endpoints related to disease management
- name: symptom
  description: Endpoints related to symptom management
paths:
  /disease:
    put:
      tags:
        - disease
      summary: Creates a disease
      operationId: create_plant_disease
      description: No restrictions.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Disease'
        required: true
      responses:
        200:
          description: Disease created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DiseaseHiddenDetails'
                  - $ref: '#/components/schemas/Disease'
        422:
          description: Group name already exists
          content: {}
  /disease/{disease_id}:
    put:
      tags:
        - disease
      summary: Gets information about the requested disease
      operationId: get_disease_info
      description: No restrictions.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Disease'
        required: true
      responses:
        200:
          description: Disease created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DiseaseHiddenDetails'
                  - $ref: '#/components/schemas/Disease'
        422:
          description: Group name already exists
          content: {}

components:
  schemas:
    DiseaseID:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: Internal ID of the group, as seen in the database
          format: int64
    Disease:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the disease
          minLength: 3
          maxLength: 128
        scientific_name:
          type: string
          description: Scientific name of the causal agent
        fruit:
          type: string
          description: Name of the fruit in which the disease is known to occur
        natural_solution:
          type: string
          description: Natural solution to get rid of the disease / natural preventive measures
        chemical_solution:
          type: string
          description: Chemical solution to get rid of the disease / chemical preventive measures
        external_link:
          type: string
          format: uri
          description: Link to buy the chemical solution
    DiseaseHiddenDetails:
      type: object
      properties:
        id:
          type: integer
          description: Internal ID of the group, as seen in the database
          format: int64
    GroupArray:
      type: object
      required:
        - name
      properties:
        groups:
          type: array
          description: An array of group objects
          items:
            $ref: '#/components/schemas/Group'


  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT